# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "opendataAPI",
        "description": "Open data API from open dataset",
        "contact": {
            "email": "mellyniaramadhan762@gmail.com"
        },
        "license": {
            "name": "mellyniaramadhan",
            "url": "http://your-license-url.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/Kominfo/UmkmGoOnlines": {
            "get": {
                "tags": [
                    "Umkm Go Onlines"
                ],
                "summary": "Get a paginated list of Umkm Go Onlines data from Kominfo.",
                "operationId": "31004035f2627c51124f3f662f322980",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UmkmGoOnlines"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/Kominfo": {
            "post": {
                "tags": [
                    "Umkm Go Onlines"
                ],
                "summary": "Create a new Umkm Go Online record.",
                "operationId": "9d6493ff9bf3c0933f833fa013b4c792",
                "requestBody": {
                    "description": "Data to be added",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "LocationOfMarketGrebegActivities": {
                                        "type": "string"
                                    },
                                    "OnBoardingAchievments": {
                                        "type": "ineteger"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "LocationOfMarketGrebegActivities": "example LocationOfMarketGrebegActivities",
                                "OnBoardingAchievments": "example OnBoardingAchievments"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "LocationOfMarketGrebegActivities": {
                                            "type": "string",
                                            "example": "Kota Jakarta"
                                        },
                                        "OnBoardingAchievments": {
                                            "type": "integer",
                                            "example": "2011"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "fieldName": [
                                                    "Error message"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Kominfo/UmkmGoOnlines/{id}": {
            "get": {
                "tags": [
                    "Umkm Go Onlines"
                ],
                "summary": "Get a specific Umkm Go Onlines record by ID.",
                "description": "Display the specified resource.",
                "operationId": "a6aee0248c9e5a4b331aa139cafc83ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Umkm Go Onlines record to be retrieved.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UmkmGoOnlines"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Message": {
                                            "type": "string",
                                            "example": "Not Found!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Umkm Go Onlines"
                ],
                "summary": "Update Umkm Go Onlines data by districts",
                "description": "Update Umkm Go Onlines data based on the given ID",
                "operationId": "updateUmkmGoOnlines",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Umkm Go Onlines data to be updated",
                        "required": true,
                        "schema": {
                            "properties": {
                                "LocationOfMarketGrebegActivities": {
                                    "type": "string"
                                },
                                "OnBoardingAchievments": {
                                    "type": "integer"
                                }
                            },
                            "type": "object",
                            "example": {
                                "LocationOfMarketGrebegActivities": "example LocationOfMarketGrebegActivities",
                                "OnBoardingAchievments": "example OnBoardingAchievments"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "description": "Umk Go Onlines data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Umkm Go Onlines data updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "LocationOfMarketGrebegActivities": {
                                            "description": "Location Of Market Grebeg Activities",
                                            "type": "string",
                                            "example": "Kota Jakarta"
                                        },
                                        "OnBoardingAchievments": {
                                            "description": "On Boarding Achievments",
                                            "type": "integer",
                                            "example": "2011"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Umkm Go Onlines data not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Umkm Go Onlines"
                ],
                "summary": "Delete a Umkm Go Onlines record by ID.",
                "operationId": "ea4fab3021ee8914ba8f38bde8224a7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Umkm Go Onlines record to be deleted.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Message": {
                                            "type": "string",
                                            "example": "Deleted Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UmkmGoOnlines": {
                "properties": {
                    "LocationOfMarketGrebegActivities": {
                        "type": "string"
                    },
                    "OnBoardingAchievments": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    }
}